requirements.txt
#* Core
Django==4.1.5
djangorestframework==3.14.0
coreapi==2.3.3
openapi-codec==1.3.2
simplejson==3.18.1

#* Docs
mkdocs==1.4.2